all: add v2 _content concurrently with existing _content
This change creates the ability to have concurrent versions of
the website running so that both websites can run concurrently
and then eventually after full on approvals, we can delete the
incumbent code in _content and rename _content_v2 to _content.

Essentially our redesign now lives in "_content_v2/" and the
current website can continually be developed as is in "_content".